By the following reason these generated files are brought under version control.
First of all, these image are necessary for the correct compilation of the
np-complete article. There they should be present in the final output.
Second of all, an export of this article should be self contained. By the 
above argument these files should be under version control.
Lastly, It is possible that these files will never change again, but should
they change, the changes reflect interesting change of view which will be
visible in the revision history.