Created a new snippet.
The new snippet in the create-article script contains a set of often needed
environments. The environments created are:
notation, definition, remark, example, conjecture, lemma, theorem,
proposition, corollary and last but not least proof.

Furthermore, a mechanism to create more environments with the same style is
given. This is a convenience method to produce a large number of similiar
looking environments.